nautilus extension
I add two extensions for nautilus to interact with the Nextcloud linux client, For the first one, I add a right click direct access for access to Copy and Move to Nextcloud. This will directly copy and move selected files to our server. The second extension its a shareable link access in the Nautilus file explorer. This will allow users to automatically create links from the file explorer without the need of going to the website to do it.
Then I created a script that once executed will create a folder in our home directory where our files should be located, paste those files on there, install the necessary  dependencies and reboot Nautilus file explorer. This script will be executed when the CMake file inside the nautilus folder install.